### 2020-09-10 aff3dd5
Increase coverage to > 90%, added Categorical test
