+++ test/CategoricalWalk1D.jl
@@ -1,0 +1,142 @@
+# using Revise
+using POMDPStressTesting
+using Distributions
+using Parameters
+
+
+@with_kw mutable struct CategoricalWalk1DParams
+    startx::Float64 = 0 # Starting x-position
+    threshx::Float64 = 10 # +- boundary threshold
+    endtime::Int64 = 30 # Simulate end time
+end
+
+
+# Implement abstract GrayBox.Simulation
+@with_kw mutable struct CategoricalWalk1DSim <: GrayBox.Simulation
+    params::CategoricalWalk1DParams = CategoricalWalk1DParams() # Parameters
+    x::Float64 = 0 # Current x-position
+    t::Int64 = 0 # Current time
+    distribution::Distribution = Categorical([0.95,0.025,0.025]) # Transition distribution [1,2,3]
+end
+
+
+# Override from GrayBox
+GrayBox.environment(sim::CategoricalWalk1DSim) = GrayBox.Environment(:x => sim.distribution)
+
+
+# Override from GrayBox (NOTE: used with ASTSeedAction)
+function GrayBox.transition!(sim::CategoricalWalk1DSim)
+    # We sample the environment and apply the transition
+    environment::GrayBox.Environment = GrayBox.environment(sim) # Get the environment distributions
+    sample::GrayBox.EnvironmentSample = rand(environment) # Sample from the environment
+    sim.t += 1 # Keep track of time
+    sim.x += sample[:x].value # Move agent using sampled value from input
+    return logpdf(sample)::Real # Summation handled by `logpdf()`
+end
+
+
+# Override from GrayBox (NOTE: used with ASTSampleAction)
+function GrayBox.transition!(sim::CategoricalWalk1DSim, sample::GrayBox.EnvironmentSample)
+    # The environment was sampled for us, and we just apply the transition
+    sim.t += 1 # Keep track of time
+    sim.x += sample[:x].value # Move agent using sampled value from input
+    return logpdf(sample)::Real # Summation handled by `logpdf()`
+end
+
+
+# Override from BlackBox
+function BlackBox.initialize!(sim::CategoricalWalk1DSim)
+    sim.t = 0
+    sim.x = sim.params.startx
+end
+
+
+# Override from BlackBox
+BlackBox.distance(sim::CategoricalWalk1DSim) = max(sim.params.threshx - abs(sim.x), 0)
+
+
+# Override from BlackBox
+BlackBox.isevent(sim::CategoricalWalk1DSim) = abs(sim.x) >= sim.params.threshx
+
+
+# Override from BlackBox
+BlackBox.isterminal(sim::CategoricalWalk1DSim) = BlackBox.isevent(sim) || sim.t >= sim.params.endtime
+
+
+# Override from BlackBox (NOTE: used with ASTSeedAction)
+function BlackBox.evaluate!(sim::CategoricalWalk1DSim)
+    logprob::Real = GrayBox.transition!(sim) # Step simulation
+    d::Real = BlackBox.distance(sim) # Calculate miss distance
+    event::Bool = BlackBox.isevent(sim) # Check event indication
+    return (logprob::Real, d::Real, event::Bool)
+end
+
+
+# Override from BlackBox (NOTE: used with ASTSampleAction)
+function BlackBox.evaluate!(sim::CategoricalWalk1DSim, sample::GrayBox.EnvironmentSample)
+    logprob::Real = GrayBox.transition!(sim, sample) # Step simulation given input sample
+    d::Real = BlackBox.distance(sim) # Calculate miss distance
+    event::Bool = BlackBox.isevent(sim) # Check event indication
+    return (logprob::Real, d::Real, event::Bool)
+end
+
+
[…]
