+++ test/runtests.jl
@@ -39,6 +39,7 @@ end
 @test begin
     @info "Extra functions test"
     (planner, _, _) = run_ast(solver=MCTSPWSolver)
+    d3tree = visualize(planner)
     action_trace = search!(planner; verbose=true)
     actions = online_path(planner.mdp, planner)
     x_trace = playback(planner, actions, sim->sim.x; return_trace=true)
@@ -56,6 +57,10 @@ end
     solver.resume = true
     search!(planner)
 
+    # No top_k action trace
+    solver.resume = false
+    planner.mdp.params.top_k = 0
+    ast_action = search!(planner)
     true
 end
 
@@ -75,5 +80,23 @@ test_solvers(skip_trpo=true) # TRPO slows down the entire test suite, so skip it
 @test begin
     @info "EpisodicWalk1D"
     include("EpisodicWalk1D.jl")
+    true
+end
+
+
+@test begin
+    @info "CategoricalWalk1D"
+    include("CategoricalWalk1D.jl")
+
+    # Policy saving
+    (planner, _, _) = run_ast(solver=PPOSolver)
+    solver = PPOSolver(num_episodes=100, episode_length=30, save=true, verbose=true, policy_type=:discrete)
+    planner = solve(solver, planner.mdp)
+    search!(planner)
+
+    # Policy loading
+    solver.resume = true
+    search!(planner)
+
     true
 end
