### 2020-09-12 b45e83c
Docstring of BlackBox.evaluate! output Tuple
