### 2020-09-16 #2655 @bstabler
@danielskatz - I am interested in learning more about Julia.  However, I just read the paper and I don't think it's a good fit for my expertise either.  I can shepherd it along though if needed.  Thanks.
