### 2020-09-16 #2655 @sbenthall
This one is up my alley. I can edit it.
