### 2020-09-25 8804a14
Simplified installation since MCTS/CEM in General and added Julia v1.5 req
