+++ .travis.yml
@@ -5,9 +5,6 @@ os:
   - osx
 
 julia:
-  - 1.2
-  - 1.3
-  - 1.4
   - 1.5
 
 notifications:
