### 2020-10-01 fefaea0
Use of Flux 0.10+ and Zygote, relax Julia version
