+++ README.md
@@ -63,10 +63,10 @@ An example implementation of the AST interface is provided for the Walk1D proble
 
 # Installation
 
-Install the POMDPs.jl registry then the `POMDPStressTesting.jl` package via:
+Install `RLInterface.jl` then the `POMDPStressTesting.jl` package via:
 ```julia
 using Pkg
-pkg"registry add https://github.com/JuliaPOMDP/Registry"
+pkg"add https://github.com/JuliaPOMDP/RLInterface.jl"
 pkg"add https://github.com/sisl/POMDPStressTesting.jl"
 ```
 

[…]
