### 2020-10-01 09a0bbc
Travis CI removed unnecessary General registry
