### 2020-10-01 e378f7c
Add support for Julia v1.0
