+++ .travis.yml
@@ -6,7 +6,7 @@ os:
 
 julia:
   - 1.0
-  - 1.3
+  - 1.2
   - 1.5
 
 notifications:
@@ -19,7 +19,7 @@ addons:
       - python3-matplotlib
 
 script:
-  - julia -e 'using Pkg; pkg"add https://github.com/JuliaPOMDP/RLInterface.jl"'
+  - julia test/deps.jl
   - julia -e 'using Pkg; Pkg.build(); Pkg.test(coverage=true)'
 after_success:
   - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(process_folder()); Coveralls.submit(process_folder())'
