### 2020-10-01 47fade6
Travis: added test/deps.jl for Julia 1.0 support
