### 2020-10-01 0eaf69b
Travis Julia 1.0: "dev" POMDPs and RLInterface at the same time
