+++ test/deps.jl
@@ -1,8 +1,7 @@
 using Pkg
 
 if VERSION < v"1.1"
-    pkg"dev https://github.com/JuliaPOMDP/POMDPs.jl"
-    pkg"dev https://github.com/JuliaPOMDP/RLInterface.jl"
+    pkg"dev https://github.com/JuliaPOMDP/POMDPs.jl https://github.com/JuliaPOMDP/RLInterface.jl"
 else
     pkg"add https://github.com/JuliaPOMDP/RLInterface.jl"
 end
