### 2020-10-06 #2655 @jarvist
> @jarvist would you be able to review this article?

Having looked quickly at the package, my expertise is close enough that I could review it. With the start of term I'm also quite busy, so it may take a few weeks to schedule in.
If that's an OK timetable, then please do assign me as a reviewer!
