+++ notebooks/walk1d.jl
@@ -1,0 +1,301 @@
+### A Pluto.jl notebook ###
+# v0.11.14
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 92ce9460-f62b-11ea-1a8c-179776b5a0b4
+using Distributions, Parameters, POMDPStressTesting, Latexify
+
+# ╔═╡ 2978b840-f62d-11ea-2ea0-19d7857208b1
+md"""
+# Black-Box Stress Testing
+"""
+
+# ╔═╡ 1e00230e-f630-11ea-3e40-bf8c852f78b8
+# begin
+#   using Pkg
+#   pkg"registry add https://github.com/JuliaPOMDP/Registry"
+#   pkg"add https://github.com/sisl/RLInterface.jl"
+#   pkg"add https://github.com/sisl/POMDPStressTesting.jl"
+# end
+md"*Unhide for installation (waiting on Julia registry).*"
+
+# ╔═╡ 40d3b1e0-f630-11ea-2160-01338d9f2209
+md"""
+To find failures in a black-box autonomous system, we can use the `POMDPStressTesting` package which is part of the POMDPs.jl ecosystem.
+
+Various solvers—which adhere to the POMDPs.jl interface—can be used:
+- `MCTSPWSolver` (MCTS with action progressive widening)
+- `TRPOSolver` and `PPOSolver` (deep reinforcement learning policy optimization)
+- `CEMSolver` (cross-entropy method)
+- `RandomSearchSolver`
+"""
+
+# ╔═╡ 86f13f60-f62d-11ea-3241-f3f1ffe37d7a
+md"""
+### Simple Problem: One-Dimensional Walk (Walk$1$D)
+We define a simple problem for adaptive stress testing (AST) to find failures. This problem, called Walk1D, samples random walking distances from a standard normal distribution $\mathcal{N}(0,1)$ and defines failures as walking past a certain threshold (which is set to $\pm 10$ in this example). AST will either select the seed which deterministically controls the sampled value from the distribution (i.e. from the transition model) or will directly sample the provided environmental distributions. These action modes are determined by the seed-action or sample-action options. AST will guide the simulation to failure events using a notion of distance to failure, while simultaneously trying to find the set of actions that maximizes the log-likelihood of the samples.
+"""
+
+# ╔═╡ d3411dd0-f62e-11ea-27d7-1b2ed8edc415
+md"""
+### Gray-Box Simulator and Environment
+The simulator and environment are treated as gray-box because we need access to the state-transition distributions and their associated likelihoods.
+"""
+
+# ╔═╡ e37d7542-f62e-11ea-0b61-513a4b44fc3c
+md"""
+##### Parameters
+First, we define the parameters of our simulation.
+"""
+
+# ╔═╡ fd7fc880-f62e-11ea-15ac-f5407aeff2a6
+@with_kw mutable struct Walk1DParams
+    startx::Float64 = 0   # Starting x-position
+    threshx::Float64 = 10 # +- boundary threshold
+    endtime::Int64 = 30   # Simulate end time
+end;
+
+# ╔═╡ 012c2eb0-f62f-11ea-1637-c113ad01b144
+md"""
+##### Simulation
[…]
