+++ notebooks/walk1d.jl
@@ -5,7 +5,7 @@ using Markdown
 using InteractiveUtils
 
 # ╔═╡ 92ce9460-f62b-11ea-1a8c-179776b5a0b4
-using Distributions, Parameters, POMDPStressTesting, Latexify
+using Distributions, Parameters, POMDPStressTesting, Latexify, PlutoUI
 
 # ╔═╡ 2978b840-f62d-11ea-2ea0-19d7857208b1
 md"""
@@ -34,13 +34,13 @@ Various solvers—which adhere to the POMDPs.jl interface—can be used:
 
 # ╔═╡ 86f13f60-f62d-11ea-3241-f3f1ffe37d7a
 md"""
-### Simple Problem: One-Dimensional Walk (Walk$1$D)
+## Simple Problem: One-Dimensional Walk
 We define a simple problem for adaptive stress testing (AST) to find failures. This problem, called Walk1D, samples random walking distances from a standard normal distribution $\mathcal{N}(0,1)$ and defines failures as walking past a certain threshold (which is set to $\pm 10$ in this example). AST will either select the seed which deterministically controls the sampled value from the distribution (i.e. from the transition model) or will directly sample the provided environmental distributions. These action modes are determined by the seed-action or sample-action options. AST will guide the simulation to failure events using a notion of distance to failure, while simultaneously trying to find the set of actions that maximizes the log-likelihood of the samples.
 """
 
 # ╔═╡ d3411dd0-f62e-11ea-27d7-1b2ed8edc415
 md"""
-### Gray-Box Simulator and Environment
+## Gray-Box Simulator and Environment
 The simulator and environment are treated as gray-box because we need access to the state-transition distributions and their associated likelihoods.
 """
 
@@ -73,7 +73,7 @@ end;
 
 # ╔═╡ 11e445d0-f62f-11ea-305c-495272981112
 md"""
-#### GrayBox.environment
+### GrayBox.environment
 Then, we define our `GrayBox.Environment` distributions. When using the `ASTSampleAction`, as opposed to `ASTSeedAction`, we need to provide access to the sampleable environment.
 """
 
@@ -82,7 +82,7 @@ GrayBox.environment(sim::Walk1DSim) = GrayBox.Environment(:x => sim.distribution
 
 # ╔═╡ 48a5e970-f62f-11ea-111d-35694f3994b4
 md"""
-#### GrayBox.transition!
+### GrayBox.transition!
 We override the `transition!` function from the `GrayBox` interface, which takes an environment sample as input. We apply the sample in our simulator, and return the log-likelihood.
 """
 
@@ -95,13 +95,13 @@ end
 
 # ╔═╡ 6e111310-f62f-11ea-33cf-b5e943b2f088
 md"""
-### Black-Box System
+## Black-Box System
 The system under test, in this case a simple single-dimensional moving agent, is always treated as black-box. The following interface functions are overridden to minimally interact with the system, and use outputs from the system to determine failure event indications and distance metrics.
 """
 
 # ╔═╡ 7c84df7e-f62f-11ea-3b5f-8b090654df19
 md"""
-#### BlackBox.initialize!
+### BlackBox.initialize!
 Now we override the `BlackBox` interface, starting with the function that initializes the simulation object. Interface functions ending in `!` may modify the `sim` object in place.
 """
 
[…]
