### 2020-10-14 #2749 @whedon
[REVIEW]: POMDPStressTesting.jl: Adaptive Stress Testing for Black-Box Systems
**Submitting author:** @mossr (<a href="http://orcid.org/0000-0003-2403-454X">Robert Moss</a>)
**Repository:** <a href="https://github.com/sisl/POMDPStressTesting.jl" target ="_blank">https://github.com/sisl/POMDPStressTesting.jl</a>
**Version:** v0.5.0
