### 2020-11-06 #2749 @sambitdash
On the software documentation, a lot is lacking. It's not very clear what is the system all about. Why should someone use the library? It assumes prior knowledge but that cannot be expected of a software. If there is a prereq to be read before working on the library then that should be explained. There is a sample PDF file given to try the example, but generally a lot is not clear to me from the level of documentation given. Most likely I may not use the package if I just read the documentation unless I look into the algorithms and try a meaningful search on them to understand.

It's completely a subordinate system to POMDPs.jl. So explanation of the system may be needed. Why a graybox or a blackbox interface needs to be explained as well. Once, the documentation is organized this will be a great addition to the Julia ecosystem.
