### 2020-11-06 #2749 @mossr
Hey @sambitdash and @sbenthall, thank you both for your input! I agree, documentation should be centralized. I've already started to write formal documentation and will have a completed version up this afternoon.
