+++ src/BlackBox.jl
@@ -1,5 +1,5 @@
 """
-Provides virtual interface for the black-box system under test (SUT)
+Provides virtual interface for the black-box system under test (SUT).
 """
 module BlackBox
 
@@ -25,7 +25,7 @@ function initialize!(sim::GrayBox.Simulation)::Nothing end
     evaluate!(sim::GrayBox.Simulation, sample::GrayBox.EnvironmentSample)::Tuple{logprob::Real, miss_distance::Real, isevent::Bool}
 
 Evaluate the SUT given some input seed and current state, returns `logprob`, `miss_distance`, and `isevent` indication.
-If the `sample` version is implemented, then ASTSampleAction will be used instead of ASTSeedAction.
+If the `sample` version is implemented, then `ASTSampleAction` will be used instead of `ASTSeedAction`.
 """
 function evaluate!(sim::GrayBox.Simulation)::Tuple{Real, Real, Bool} end
 function evaluate!(sim::GrayBox.Simulation, sample::GrayBox.EnvironmentSample)::Tuple{Real, Real, Bool} end
@@ -43,7 +43,7 @@ function distance(sim::GrayBox.Simulation)::Real end
 """
     isevent(sim::GrayBox.Simulation)::Bool
 
-Return a boolean indicating if the SUT reached an event of interest.
+Return a boolean indicating if the SUT reached a failure event of interest.
 """
 function isevent(sim::GrayBox.Simulation)::Bool end
 
