+++ docs/make.jl
@@ -1,0 +1,19 @@
+using Documenter, POMDPStressTesting
+
+makedocs(
+    modules = [POMDPStressTesting, AST, BlackBox, GrayBox],
+    sitename = "POMDPStressTesting.jl",
+    authors = "Robert Moss",
+    clean = false,
+    pages = [
+        "Home" => "index.md",
+        "Manual" => [
+            "Installation" => "install.md",
+            "Guide" => "guide.md",
+            "Solvers" => "solvers.md",
+        ],
+        "Example" => "example.md",
+        "Library/Interface" => "interface.md",
+        "Contributing" => "contrib.md"
+    ]
+)
