+++ .travis.yml
@@ -29,7 +29,7 @@ jobs:
     - os: windows
   include:
     - stage: "Documentation"
-      julia: 1
+      julia: 1.5
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
