### 2020-11-06 527f1a2
Travis CI docs to develop in current directory
