+++ docs/make.jl
@@ -1,5 +1,7 @@
 using Documenter, POMDPStressTesting
 
+# TODO: Literate.jl to import notebook
+
 makedocs(
     modules = [POMDPStressTesting, AST, BlackBox, GrayBox],
     sitename = "POMDPStressTesting.jl",
