+++ .travis.yml
@@ -2,11 +2,11 @@
 language: julia
 os:
   - windows
-  - osx
+  # - osx
 
 julia:
-  - 1.0
-  - 1.2
+  # - 1.0
+  # - 1.2
   - 1.5
 
 notifications:
@@ -33,6 +33,6 @@ jobs:
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
-                                    Pkg.develop(PackageSpec(path=pwd()))'
+                                    pkg"dev ."'
         - julia --project=docs/ docs/make.jl
       after_success: skip
