### 2020-11-06 a4864b8
Travis CI docs to add packages not in registry
