+++ .travis.yml
@@ -32,7 +32,10 @@ jobs:
       julia: 1.5
       os: linux
       script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
-                                    pkg"dev ."'
+        - julia --project=docs/ -e 'using Pkg;
+                                    pkg"add https://github.com/JuliaPOMDP/RLInterface.jl";
+                                    pkg"add https://github.com/sisl/POMDPStressTesting.jl";
+                                    Pkg.instantiate();
+                                    Pkg.develop(PackageSpec(path=pwd()))'
         - julia --project=docs/ docs/make.jl
       after_success: skip
