### 2020-11-06 d4b4236
Travis CI docs on Windows
