+++ .travis.yml
@@ -29,6 +29,8 @@ jobs:
     - stage: "Documentation"
       julia: 1.5
       os: windows
+      env:
+        DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
       script:
         - julia --project=docs/ -e 'using Pkg;
                                     pkg"add https://github.com/JuliaPOMDP/RLInterface.jl";
