### 2020-11-06 13e36e1
Travis CI docs added GITHUB_TOKEN
