+++ test/runtests.jl
@@ -1,10 +1,6 @@
 using Test
 using NBInclude
 
-@test true # TEMPORARY FOR DOCUMENTATION PURPOSES
-
-#=
-
 testdir = joinpath(dirname(@__DIR__), "test")
 notebookdir = joinpath(dirname(@__DIR__), "notebooks")
 
@@ -104,5 +100,3 @@ end
 
     true
 end
-
-=#
