+++ .travis.yml
@@ -2,11 +2,11 @@
 language: julia
 os:
   - windows
-  # - osx
+  - osx
 
 julia:
-  # - 1.0
-  # - 1.2
+  - 1.0
+  - 1.2
   - 1.5
 
 notifications:
