+++ .travis.yml
@@ -5,7 +5,7 @@ os:
   - osx
 
 julia:
-  - 1.0
+  # - 1.0
   - 1.2
   - 1.5
 
