### 2020-11-07 #2749 @mossr
@sambitdash I've pushed comprehensive documentation that addresses the concerns you raised and checks all the necessary boxes above.

https://sisl.github.io/POMDPStressTesting.jl/dev/

Thank you again for the constructive input! Let me know if you have any more questions or would like additional clarifications.
