### 2020-11-12 d65e860
Added `pass_seed_action` to ASTParams
