+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"guide/#guide","page":"Guide","title":"Guide","text":"","category":"section"},{"location":"guide/","page":"Guide","title":"Guide","text":"To use this package to stress test your own system, the user has to provide the following:","category":"page"},{"location":"guide/","page":"Guide","title":"Guide","text":"Implementations of the BlackBox system interface functions (outlined here)\nImplementations of the GrayBox simulator/environment interface functions (outlined here)","category":"page"},{"location":"guide/#Problem-Setup","page":"Guide","title":"Problem Setup","text":"","category":"section"},{"location":"guide/","page":"Guide","title":"Guide","text":"Once the GrayBox and BlackBox interfaces are defined (see Example for a full example), the user has access to the solvers and simply needs to set up the AST problem.","category":"page"},{"location":"guide/","page":"Guide","title":"Guide","text":"First, set up your simulation structure, where YourSimulation <: GrayBox.Simulation (note, you'll need to replace YourSimulation with your own structure you've defined as part of the GrayBox interface):","category":"page"},{"location":"guide/","page":"Guide","title":"Guide","text":"sim::GrayBox.Simulator = YourSimulation()","category":"page"},{"location":"guide/","page":"Guide","title":"Guide","text":"Then, set up the adaptive stress testing (AST) Markov decision process (MDP) structure, given your sim object:","category":"page"},{"location":"guide/","page":"Guide","title":"Guide","text":"mdp::ASTMDP = ASTMDP{ASTSeedAction}(sim) # could use `ASTSeedAction` or `ASTSampleAction`","category":"page"},{"location":"guide/#ast_action_type","page":"Guide","title":"AST Action Type","text":"","category":"section"},{"location":"guide/","page":"Guide","title":"Guide","text":"ASTSeedAction\nThis action type samples seeds for a random number generator (RNG), which means the GrayBox.transition! function must sample from the environment themselves and apply the transition.\nUseful when it's difficult to access the individual sampled environment outputs.\nASTSampleAction\nThis action type samples directory from the GrayBox.Environment and will pass the sample(s) (as GrayBox.EnvironmentSample) to the GrayBox.transition! function to be directly applied\nUseful when you have full access to the simulation environment and can apply each sample directly in the transition function.","category":"page"},{"location":"guide/#Solving-the-AST-MDP","page":"Guide","title":"Solving the AST MDP","text":"","category":"section"},{"location":"guide/","page":"Guide","title":"Guide","text":"Now you can choose your solver (see Solvers) and run solve given the AST mdp (Markov decision process) to produce an online planner (no search has been performed yet):","category":"page"},{"location":"guide/","page":"Guide","title":"Guide","text":"solver = MCTSPWSolver()\nplanner = solve(solver, mdp)","category":"page"},{"location":"guide/#Searching-for-Failures","page":"Guide","title":"Searching for Failures","text":"","category":"section"},{"location":"guide/","page":"Guide","title":"Guide","text":"Once the problem is set up, you can search for failures using search! given your planner. This will return the best action trace it found.","category":"page"},{"location":"guide/","page":"Guide","title":"Guide","text":"action_trace = search!(planner)","category":"page"},{"location":"guide/#Playback-Failures","page":"Guide","title":"Playback Failures","text":"","category":"section"},{"loc
[…]
