### 2020-11-12 #2749 @jarvist
> @sambitdash I've pushed comprehensive documentation that addresses the concerns you raised and checks all the necessary boxes above.

This is much better!
I think it could be improved still further though - it's probably worth expanding the POMDPS acronym, and explaining how this technique is used; and also defining what a 'black-box' system is in this context; and for that matter what counts as 'failure'.

Can the Walk1D example be simplified any further? It's a bit confusing that the first thing the notebook starts talking about is how the problem is set up as a 'Gray box'.
