### 2020-11-12 f9aeaa8
POMDP definition, failure/distance definition, and CEM conversion function
