+++ docs/src/index.md
@@ -4,6 +4,8 @@
 This package is used to find likely failures in a black-box software system.
 The package is integrated with the [POMDPs.jl](https://github.com/JuliaPOMDP/POMDPs.jl) ecosystem; giving access to solvers, policies, and visualizations (although no prior knowledge of the POMDPs.jl package is needed to use POMDPStressTesting.jl—see the [Guide](@ref guide)). It uses a technique called adaptive stress testing (AST)[^1] to find likely failures using a distance metric to a failure event and the likelihood of an environment sample to guide the search.
 
+A POMDP is a [partially observable Markov decision process](https://en.wikipedia.org/wiki/Partially_observable_Markov_decision_process), which is a framework to define a sequential decision making problem where the true state is unobservable. In the context of this package, we use the POMDP acronym mainly to tie the package to the POMDPs.jl package, but the system that is stress tested can also be defined as a POMDP.
+
 This package is intended to help developers stress test their systems before deployment into the real-world (see existing use cases for aircraft collision avoidance systems[^1] and aircraft trajectory prediction systems[^2]). It is also used for research purposes to expand on the AST concept by allowing additional solution methods to be explored and tested.
 
 ---
@@ -56,6 +58,11 @@ The `GrayBox` simulator and environment interface includes:
 - `GrayBox.Simulation` type to hold simulation variables
 - `GrayBox.environment(sim::Simulation)` to return the collection of environment distributions
 - `GrayBox.transition!(sim::Simulation)` to transition the simulator, returning the log-likelihood
+
+## Failure and Distance Definition
+A *failure* event of the system under test is defined be the user. The user defines the function `BlackBox.isevent` to return an boolean indicating a failure or not given the current state of the simulation. An example failure used in the context of AST would be a collision when stress testing autonomous vehicles or aircraft collision avoidance systems.
+
+The real-valued *distance* metric is used to indicate "how close are we to a failure?" and is defined by the user in the `BlackBox.distance` function. This metric is used to guide the search process towards failures by receiving a signal of the distance to a failure. An example distance metric for the autonomous vehicle problem would be the distance between the autonomous vehicle and a pedestrian, where if a failure is a collision with a pedestrian then we'd like to minimize this distance metric to find failures.
 
 ## Citation
 
