### 2020-11-12 8bd48e2
CEM `search!` function split
