### 2020-11-13 #2749 @mossr
@jarvist I agree with your comments—thank you for these. I've since added a paragraph describing the POMDP acronym and why we use it, and added a section on the definition of "failure" and the "distance" metric, all under the "Home" section here:
- **Home**: https://sisl.github.io/POMDPStressTesting.jl/dev/

Note that I already have two sections describing the "black-box" and "gray-box" definitions, again under "Home".

The Walk1D example *is* the simplest example we could come up with, but I've since included references back to the gray-box/black-box sections in the documentation that help describe these components. Also, I'd refer you to the "Example" section of the documentation which helps define the Walk1D problem definitions (i.e. the system/environment/failure/distance), which I also included in the notebook to centralize that information.
- **Example**: https://sisl.github.io/POMDPStressTesting.jl/dev/example/
- **Notebook**: https://nbviewer.jupyter.org/github/sisl/POMDPStressTesting.jl/blob/master/notebooks/Walk1D.ipynb
