### 2020-11-13 #2749 @jarvist
Dear @mossr , that's great! It's extremely clear now, and I think would answer all the questions that arose when I first read through it. The lay-person definition followed by an example is very easy to follow. I'm certainly happy with this documentation now.

Apologies for the sluggishness of doing my review, I'll install the software, check the repo & read the paper as soon as I can fit it in.
