### 2021-02-09 #2749 @jarvist
@mossr are the fairly heavy GUI requirements of the package (Seaborn, Pyplot I presume) necessary? They are currently causing a failure for me at the 'Precompiling POMDPStressTesting' just after using, I think due to some automatic Python compile failing on the backend. Generally these requirements can make the code fragile for installation, and be quite limiting in some HPC environments (e.g. when no X is installed).

Is plotting like this integral to the package, or could it be provided separately / via Plot recipes? (https://docs.juliaplots.org/latest/recipes/)
