### 2021-02-09 #2749 @mossr
@jarvist you're absolutely right—the plotting should be separate and is not vital to the core package. I'll work on something right now to bypass the plotting unless you already load in those packaged. Hang tight.
