### 2021-02-09 e5bc942
Removed explicit dependency on PyPlot/Seaborn (now handled by Requires.jl)
