+++ Project.toml
@@ -19,10 +19,9 @@ POMDPSimulators = "e0d0a172-29c6-5d4e-96d0-f262df5d01fd"
 POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
-PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 RLInterface = "3065fe53-3878-5aed-aae7-ab641be43734"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Seaborn = "d2ef9438-c967-53ab-8060-373fdd9e13eb"
+Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
@@ -36,7 +35,8 @@ julia = "1"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 NBInclude = "0db19996-df87-5ea3-a455-e3a50d440464"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Distributions", "NBInclude", "Parameters", "Test"]
+test = ["Distributions", "NBInclude", "Parameters", "Pkg", "Test"]
