### 2021-02-09 #2749 @mossr
@jarvist I relaxed the PyPlot/Seaborn plotting dependencies using Requires.jl (and subsequently updated the example Jupyter notebook and documentation to reflect this design choice). This is tagged in a new release v0.4.4 (master).

Thank you for the heads up—I agree that the dependency on Python for plotting would be an unnecessary (and unfortunate) bottleneck for many applications of this package.
