### 2021-02-26 bdaf511
Trigger plot testing only on Travis, not through ]test
