+++ .travis.yml
@@ -20,7 +20,7 @@ addons:
 
 script:
   - julia test/deps.jl
-  - julia -e 'using Pkg; Pkg.build(); Pkg.test(coverage=true)'
+  - julia -e 'ENV["TESTPLOTS"]=true; using Pkg; Pkg.build(); Pkg.test(coverage=true)'
 after_success:
   - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(process_folder()); Coveralls.submit(process_folder())'
 
