### 2021-02-26 #2749 @mossr
@sbenthall I just made it _even easier_ for you to test the entire package, avoiding the plotting. So when you do start testing, pull the latest v0.4.5 (on master) and this will make your life much easier :)

Thanks again for getting around to this. Let me know if you run into any hiccups.
