### 2021-03-17 sisl/POMDPStressTesting.jl#7 @sambitdash
Please fix the 2 doc missing elements
Missing docstring.
Missing docstring for episodic_figures. Check Documenter's build log for details.

Missing docstring.
Missing docstring for distribution_figures. Check Documenter's build log for details.
