### 2021-03-17 2f23550
Call 'using PyPlot,Seaborn' during docs generation to fix #7
