### 2021-03-17 #2749 @mossr
@sambitdash fixed with https://github.com/sisl/POMDPStressTesting.jl/commit/2f23550c65148d48d70d853f3bc055f13064979c
