### 2021-03-18 #2749 @jarvist
Documentation & JOSS paper are now very nicely written ^:)
