### 2021-03-18 #2749 @jarvist
`PyPlot` is a hidden requirement for the tests to run, it appears to get called in `Walk2D`. This causes the tests to error out when halfway complete, on a new installation on Julia 1.6.0-rc2.
