### 2021-03-18 690c85e
Removed unused PyPlot call from Walk2D test
