### 2021-03-18 #2749 @mossr
@jarvist great catch! That `PyPlot` code is unused anyhow, so I've just removed it from the `Walk2D` test (updated on master).
