### 2021-03-18 #2749 @jarvist
I edited it out by hand, and then the tests ran to completion.

I'm finished with my review now - it looks like a well put together & described package! 🎉
