+++ joss/paper.bib
@@ -50,7 +50,8 @@ booktitle={{IEEE/AIAA Digital Avionics Systems Conference (DASC)}},
 title={{Adaptive Stress Testing of Trajectory Predictions in Flight Management Systems}},
 year={2020},
 volume={},
-number={}}
+number={},
+doi={10.1109/DASC50938.2020.9256730}}
 
 
 @INPROCEEDINGS{ast_av,
