### 2021-04-04 #2749 @sbenthall
OK. This looks great.

@mossr couldyou please:

- make a tagged release
- archive it so that it has a DOI
- report the version number and archive DOI in the review thread.
