### 2021-04-05 929350a
Version update to v0.5.0 (JOSS publication)
