### 2021-04-05 #2749 @mossr
@sbenthall I've tagged and released a new version v0.5.0. How would you recommend archiving it and producing a DOI?
