### 2021-04-05 #2749 @sbenthall
I recommend uploading the software bundle to Zenodo.
It should provide a DOI.
https://zenodo.org/
