### 2021-04-06 #2749 @mossr
Thank you, @sbenthall.

- Tagged release of [v0.5.0](https://github.com/sisl/POMDPStressTesting.jl/releases/tag/v0.5.0)
- Archived with a DOI of [10.5281/zenodo.4663471](https://zenodo.org/record/4663471)
