### 2021-04-08 91e95b5
removing duplicate bib entry
