+++ joss/paper.bib
@@ -150,10 +150,3 @@ doi={10.1109/IVS.2018.8500400}
   month={12},
   year={2011}
 }
-
-@article{gym,
-  title={{OpenAI} gym},
-  author={Brockman, Greg and Cheung, Vicki and Pettersson, Ludwig and Schneider, Jonas and Schulman, John and Tang, Jie and Zaremba, Wojciech},
-  journal={arXiv:1606.01540},
-  year={2016}
-}
