### 2021-04-08 #2749 @danielskatz
@mossr  - there are a few small changes in https://github.com/sisl/POMDPStressTesting.jl/pull/8 - if you can merge these, I'll go ahead and finish the acceptance and publishing
